Improved db logic

+++ Assets/Code/Data/Data.cs
@@ -1,9 +1,9 @@
 ﻿[System.Serializable]
 public class Data {
     //Kintamieji, kurie bus saugomi
-    public float level1;
-    public float level2;
-    public float level3;
+    public int level1;
+    public int level2;
+    public int level3;
     public int coins;
     public bool frogBodyOwned;
     public bool thirdPlayerBodyOwned;
